[1141] Implement new component
