+++ src/features/tasks/components/SortableTask.tsx
@@ -11,7 +11,7 @@ import { DragHandle } from "./DragHandle";
 import { Link } from "react-router";
 import { QuickDeleteTask } from "./QuickDeleteTask";
 import { ScheduledTaskWithExpectedStartTime } from "../types/scheduledTaskWithExpectedStartTime";
-import { StartTaskBtn } from "./StartTaskBtn";
+import { ScheduledTaskActionBtn } from "./ScheduledTaskActionBtn";
 
 import { formatDatetimeTo12hTime, secondsToTime } from "@/utils";
 import clsx from "clsx";
@@ -88,7 +88,7 @@ export const SortableTask = ({ task }: Props) => {
 
         <div className="flex items-center gap-3">
           <QuickDeleteTask taskId={task.id} />
-          <StartTaskBtn task={task} />
+          <ScheduledTaskActionBtn task={task} />
         </div>
       </Card>
     </div>
